refactor(theme): extrair constantes para src/constants/
- cria src/constants/theme.ts com Theme, DEFAULT_THEME, HACKER_THEME
- cria src/constants/agent.ts com AGENT_COLORS, DEFAULT_AGENT_COLOR,
  getAgentColor e AGENT_STATUS_LABEL
- ThemeContext.tsx importa de constants/theme (re-exporta por compat)
- office-layout.ts re-exporta AGENT_COLORS de constants/agent
- AgentAvatar.tsx importa AGENT_STATUS_LABEL de constants/agent

fecha #39
